+++ R/fireHazard.R
@@ -38,13 +38,32 @@
 }
 fireHazard<-function(x, SpParams, forest = NULL, standardConditions = FALSE,
                      freq="days", fun = "max") {
-  if(!inherits(x, c("spwb","pwb","growth", "fordyn"))) {
-    stop("'x' should be of class 'spwb', 'pwb', 'growth' or 'fordyn'")
+  if(!inherits(x, c("spwb","pwb","growth", "fordyn", "spwb_day", "growth_day"))) {
+    stop("'x' should be of class 'spwb', 'spwb_day', 'pwb', 'growth', 'growth_day' or 'fordyn'")
   }
   if(inherits(x, c("spwb","pwb", "growth"))) {
-    if(is.null(forest)) stop("You must supply a 'forest' object when 'x' is of class 'spwb', 'pwb' or 'growth'")
+    if(is.null(forest)) stop("You must supply a 'forest' object when 'x' is of class 'spwb', 'spwb_day', 'pwb', 'growth' or 'growth_day'")
     res = .firehaz_sim(forest, x, SpParams, standardConditions,
                        freq = freq, fun = fun)
+  } else if(inherits(x, c("spwb_day", "growth_day"))) {
+    if(is.null(forest)) stop("You must supply a 'forest' object when 'x' is of class 'spwb', 'spwb_day', 'pwb', 'growth' or 'growth_day'")
+    slope = x$topography[["slope"]]
+    weather = x$weather
+    fmc = x$Plants$LFMC
+    if(is.na(slope)) slope = 0.0
+    windSpeed = weather[["wind"]]
+    fccs = fuel_FCCS(forest, SpParams, cohortFMC = fmc)
+    vp = meteoland::utils_averageDailyVP(Tmin = weather[["tmin"]], Tmax = weather[["tmax"]],
+                                         RHmin = weather[["rhmin"]], RHmax = weather[["rhmax"]])
+    D = max(0, meteoland::utils_saturationVP(weather[["tmax"]]) - vp)
+    fm_dead = 5.43 + 52.91*exp(-0.64*D) # Resco de Dios, V., A. W. Fellows, R. H. Nolan, M. M. Boer, R. a. Bradstock, F. Domingo, and M. L. Goulden. 2015. A semi-mechanistic model for predicting the moisture content of fine litter. Agricultural and Forest Meteorology 203:64–73.
+    MdeadSI = rep(fm_dead, 5)
+    if(standardConditions) {
+      windSpeed = 11.0
+      fccs$ActFMC = NA
+      MdeadSI = c(6,6,6,6,6)
+    }
+    res = unlist(fire_FCCS(fccs, slope = slope, windSpeedSI = windSpeed, MdeadSI = MdeadSI))
   } else {
     vec<-vector("list", length(x$GrowthResults))
     for(i in 1:length(x$GrowthResults)) {