+++ NEWS.md
@@ -4,6 +4,7 @@
 
 # Version 2.7.4
 - Reducing unused parameters for Sperry model
+- Reducing computational time for Sperry model
 
 # Version 2.7.3
 - Cleaning for CRAN